+++ src/main/java/org/example/hugmeexp/domain/studyRoom/service/StudyRoomReservationService.java
@@ -21,7 +21,6 @@ import org.springframework.retry.annotation.Retryable;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.interceptor.TransactionAspectSupport;
 
 import java.time.LocalDateTime;
 import java.util.List;
